+++ core/database/src/main/java/com/same/alarm/database/mapper/AlarmMapper.kt
@@ -4,7 +4,6 @@ import com.same.alarm.data.model.AlarmEntity
 import com.same.alarm.database.model.AlarmLocal
 
 object AlarmMapper {
-
     fun AlarmLocal.toEntity(): AlarmEntity {
         return AlarmEntity(
             id = this.id,
@@ -12,7 +11,8 @@ object AlarmMapper {
             statusMessage = this.statusMessage,
             daysOfWeek = this.daysOfWeek,
             category = this.category,
-            isRepeating = this.isRepeating
+            isRepeating = this.isRepeating,
+            isActive = this.isActive
         )
     }
 
@@ -23,7 +23,8 @@ object AlarmMapper {
             statusMessage = this.statusMessage,
             daysOfWeek = this.daysOfWeek,
             category = this.category,
-            isRepeating = this.isRepeating
+            isRepeating = this.isRepeating,
+            isActive = this.isActive
         )
     }
 }
