+++ core/database/src/main/java/com/same/alarm/database/typeconverter/LocalTimeConverter.kt
@@ -1,12 +1,9 @@
 package com.same.alarm.database.typeconverter
 
-import android.os.Build
-import androidx.annotation.RequiresApi
 import androidx.room.TypeConverter
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
 
-@RequiresApi(Build.VERSION_CODES.O)
 class LocalTimeConverter {
     @TypeConverter
     fun fromLocalTime(localTime: LocalTime): String {
