--- feature/setup/src/main/java/com/same/alarm/setup/RepeatSetupViewModel.kt
@@ -1,30 +1,0 @@
-package com.same.alarm.setup
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
-import javax.inject.Inject
-
-@HiltViewModel
-class RepeatSetupViewModel @Inject constructor(
-) : ViewModel() {
-    private val _selectedDays = MutableStateFlow<List<Int>>(emptyList())
-    val selectedDays: StateFlow<List<Int>> = _selectedDays.asStateFlow()
-
-    fun setSelectedDays(initialSelectedDays: List<Int>) {
-        _selectedDays.value = initialSelectedDays
-    }
-
-    fun toggleDay(day: Int) {
-        _selectedDays.update { currentDays ->
-            if (day in currentDays) {
-                currentDays - day
-            } else {
-                (currentDays + day).sorted()
-            }
-        }
-    }
-}
