+++ feature/setup/src/main/java/com/same/alarm/setup/SetupScreen.kt
@@ -1,12 +1,14 @@
 package com.same.alarm.setup
 
-import android.util.Log
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -15,10 +17,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.same.alarm.model.Alarm
 import com.same.alarm.setup.component.CategoryDropdown
 import com.same.alarm.setup.component.ConfirmButton
@@ -30,32 +35,32 @@ import java.time.LocalTime
 
 @Composable
 fun SetupRoute(
-    setUpViewModel: SetUpViewModel = hiltViewModel(),
-    onRepeatClick: (List<Int>) -> Unit,
-    restoredSelectedDays: List<Int>,
+    setupViewModel: SetupViewModel = hiltViewModel()
 ) {
-    Log.d("okay", restoredSelectedDays.toString())
+    val selectedDays by setupViewModel.selectedDays.collectAsStateWithLifecycle()
+
     SetupScreen(
-        onRepeatClick = onRepeatClick,
-        onAddAlarm = setUpViewModel::addAlarm,
-        restoredSelectedDays = restoredSelectedDays,
+        onAddAlarm = setupViewModel::addAlarm,
+        selectedDays = selectedDays,
+        onDaySelected = setupViewModel::toggleDay
     )
 }
 
 @Composable
 fun SetupScreen(
-    onRepeatClick: (List<Int>) -> Unit,
     onAddAlarm: (Alarm) -> Unit,
-    restoredSelectedDays: List<Int>,
+    selectedDays: List<Int>,
+    onDaySelected: (Int) -> Unit
 ) {
     var time by remember { mutableStateOf(LocalTime.of(9, 0)) }
 
     var isAlarmRepeated by remember { mutableStateOf(true) }
-    var isDayOfWeekRepeated by remember { mutableStateOf(restoredSelectedDays.isNotEmpty()) }
 
     val categories = listOf("일반", "중요", "기타")
     var selectedCategory by remember { mutableStateOf(categories[0]) }
     var statusMessage by remember { mutableStateOf("") }
+
+    val days = listOf("월", "화", "수", "목", "금", "토", "일")
 
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
@@ -74,16 +79,27 @@ fun SetupScreen(
             onCheckedChange = { isAlarmRepeated = it }
         )
 
-        RepeatSwitchLabel(
-            label = "요일 반복",
-            isChecked = isDayOfWeekRepeated,
-            onCheckedChange = { isChecked ->
-                isDayOfWeekRepeated = isChecked
-                if (isChecked)  {
-                    onRepeatClick(restoredSelectedDays)
-                }
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp),
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            days.forEachIndexed { index, day ->
+                val isSelected = index in selectedDays
+
+                Text(
+                    text = day,
+                    fontSize = 16.sp,
+                    color = if (isSelected) Color.White else Color.Black,
+                    modifier = Modifier
+                        .clip(RoundedCornerShape(8.dp))
+                        .background(if (isSelected) Color.Blue else Color.LightGray)
+                        .clickable { onDaySelected(index) }
+                        .padding(8.dp)
+                )
             }
-        )
+        }
 
         Row(
             verticalAlignment = Alignment.CenterVertically,
@@ -109,21 +125,14 @@ fun SetupScreen(
             onStatusMessageChange = { statusMessage = it }
         )
 
-        Text(
-            text = restoredSelectedDays.joinToString(" "),
-            fontSize = 18.sp,
-            modifier = Modifier.weight(1f)
-        )
-
         ConfirmButton {
             onAddAlarm(
                 Alarm(
-                    0,
-                    time,
-                    statusMessage,
-                    restoredSelectedDays,
-                    selectedCategory,
-                    isAlarmRepeated
+                    time = time,
+                    statusMessage = statusMessage,
+                    daysOfWeek = selectedDays,
+                    category = selectedCategory,
+                    isRepeating = isAlarmRepeated,
                 )
             )
         }
@@ -134,8 +143,8 @@ fun SetupScreen(
 @Composable
 fun SetupScreenPreview() {
     SetupScreen(
-        onRepeatClick = {},
         onAddAlarm = {},
-        restoredSelectedDays = emptyList()
+        selectedDays = emptyList(),
+        onDaySelected = {}
     )
 }
