+++ feature/setup/src/main/java/com/same/alarm/setup/SetupViewModel.kt
@@ -5,17 +5,34 @@ import androidx.lifecycle.viewModelScope
 import com.same.alarm.domain.usecase.AddAlarmUseCase
 import com.same.alarm.model.Alarm
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class SetUpViewModel @Inject constructor(
+class SetupViewModel @Inject constructor(
     private val addAlarmUseCase: AddAlarmUseCase
 ) : ViewModel() {
 
     fun addAlarm(alarm: Alarm) {
         viewModelScope.launch {
             addAlarmUseCase(alarm)
+        }
+    }
+
+    private val _selectedDays = MutableStateFlow<List<Int>>(emptyList())
+    val selectedDays: StateFlow<List<Int>> = _selectedDays.asStateFlow()
+
+    fun toggleDay(day: Int) {
+        _selectedDays.update { currentDays ->
+            if (day in currentDays) {
+                currentDays - day
+            } else {
+                (currentDays + day).sorted()
+            }
         }
     }
 }
